changed type of facebook_user_id to string

+++ db/migrate/20111213202953_create_proposals.rb
@@ -1,12 +1,12 @@
 class CreateProposals < ActiveRecord::Migration
   def change
     create_table :proposals do |t|
-      t.bigint :user_id
+      t.integer :user_id
       t.string :groupon_id
       t.datetime :proposed_datetime
-      t.bigint :min_companions
-      t.bigint :max_companions
-      t.bigint :tipped
+      t.integer :min_companions
+      t.integer :max_companions
+      t.integer :tipped
       t.string :facebook_event_id
 
       t.timestamps
