+++ db/migrate/20111213204759_create_locations.rb
@@ -1,7 +1,7 @@
 class CreateLocations < ActiveRecord::Migration
   def change
     create_table :locations do |t|
-      t.bigint :deal_id
+      t.integer :deal_id
       t.string :street_address_1
       t.string :street_address_2
       t.string :postal_code
